+++ demo/src/main/java/com/orange/ods/demo/ui/components/ComponentVariantScreen.kt
@@ -20,6 +20,7 @@ import com.orange.ods.demo.ui.components.buttons.ComponentButtons
 import com.orange.ods.demo.ui.components.cards.ComponentCard
 import com.orange.ods.demo.ui.components.chips.Chip
 import com.orange.ods.demo.ui.components.chips.ChipFilter
+import com.orange.ods.demo.ui.components.progress.ComponentProgress
 import com.orange.ods.demo.ui.components.tabs.ComponentTabs
 import com.orange.ods.demo.ui.components.tabs.TabsConfiguration
 import com.orange.ods.demo.ui.components.textfields.ComponentTextField
@@ -43,6 +44,7 @@ fun ComponentVariantScreen(
             Component.Buttons -> ComponentButtons(variant = variant)
             Component.Cards -> ComponentCard(variant = variant)
             Component.Chips -> if (variant == Variant.ChipFilter) ChipFilter() else Chip()
+            Component.Progress -> ComponentProgress(variant = variant)
             Component.TextFields -> ComponentTextField(variant = variant)
             Component.Tabs -> ComponentTabs(variant, updateTopAppBarTabs)
             else -> {}